backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-telegram-bot[http2]==21.6
python-socketio[asyncio]==5.11.0
google-generativeai==0.8.0
python-dotenv==1.0.1
